Fixed UpdateLines, Added more features, renamed class and fixed some issues
